# bricomalin_api/public/openapi.yaml
openapi: 3.0.0
info:
  title: BricoMalin API
  version: 1.0.0
  description: API REST pour l'application BricoMalin
servers:
  - url: http://localhost:8000
    description: Serveur de développement
paths:
...
  /api/job-requests:
    get:
      summary: Liste des demandes
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des demandes
    post:
      summary: Créer une demande
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: integer
                title:
                  type: string
                description:
                  type: string
                department:
                  type: string
                city:
                  type: string
                isFree:
                  type: boolean
                suggestedPrice:
                  type: string
      responses:
        '201':
          description: Demande créée
  /api/job-requests/{id}:
    get:
      summary: Détail d'une demande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détail de la demande
...